:recycle: refactor: Separate parsers into multiple files
