+++ .editorconfig
@@ -4,3 +4,4 @@ tab_width = 2
 indent_style = space
 end_of_line = lf
 insert_final_newline = true
+charset = utf-8
